Big miscellaneous commit: implemented Actions and Inputs, TransformComponent, RenderComponent, Camera / CameraComponent, ComponentFactory, EnumFactory, started EntityAsset, debugged and fixed TEntity destructor problem.
Actions and Inputs, EnumFactory:
Completed input serialization/management within Input class. An action can be serialized as a string (such as "Left") with its joystick and keyboard mouse configuration.  Within ActionSystem, actions are mapped to inputs. Actions are not stored as strings but rather as enums, to decrease overhead. Added EnumFactory utility to refer to an enum by the string of the enum name, and vice versa. ActionEnum initializes all actions, so that the enum Left can be accessed by string "Left", and vice versa, to facilitate serialization. ActionSystem ironically is not a System (need to rename it, it stands apart from ECS), but can be solicited by any System with function: bool GetActionValue(Action).

TransformComponent / RenderComponent:
TransformComponent stores transform data which can be manipulated by systems during update, and RenderingSystem then updates sf::Sprite with TransformComponent data before drawing.
RenderComponent stores the layer number, and modified RenderingSystem add() function so that its vector of entities is sorted by lowest to greatest layer. This allows update function to effectively iterate through and draw entities by layer number, aka painters algorithm.

Camera / CameraComponent:
Camera is a system which makes sure player entity (with PlayerComponent ) is always at the center of the screen, and updates all other entitiy transforms as player moves across screen, making sure player stays in bounds. CameraComponent holds aspect ratio and bounds of camera.

EntityAsset:
Began EntityAsset with a member EntityHandle variable. this EntityHandle is a template for an entity (like a goomba) which can then be cloned. Working on serialization and Imgui Editor of EntityAsset.

ComponentFactory:
Added ComponentFactory in Entity.h which maps ComponentType string (eg. "TransformComponent") to function pointer which would add<ComponentType> to entity. Useful for ImGui debugging purposes to allow free willy adding (and eventually, removing) of ComponentTypes to an entity (accessed by string of ComponentType).

TEntity destructor problem:
Issue arose where each entity would share the same component of a ComponentType(eg.TransformComponent). When creating a component of ComponentType it would always be placed at position 0 in component_registry<ComponentType>. This is because when entities vector in EntityRegistry increased in size ~TEntity would be called, which called detach on all components, which pushed component index (0) to freeIndices<ComponentType>. Therefore when subsequently adding one of these ComponentTypes on another entity, the component would be place at the freeIndex 0. Took a while to debug.
*****To fix it, I added a function destroy() which should be called when entities are purposefully destroyed and ~TEntity is now default dtor*****