Implemented multiple collision events, detection for each collision event and resolution for static/dynamic collision, refined level editor.
Created static/dynamic, triggerstatic/dynamic, static/triggerdynamic, and dynamic/dynamic events in CollisionEvents.h. These collision events hold the two collider entities, as well as information on which side of the entities was collided with.
If an entity has a RigidBodyComponent and ColliderComponent it is dynamic. If it only has a ColliderComponent it is static. ColliderComponent holds a bool is_trigger to determine if the collision is trigger or not. A trigger collider will not undergo any physics collision resolution but an event will still be remarked and can be listened for. RigidBodyComponent also holds a bool is_kinematic.
Collision is detected in CollisionSystem with AABB and each event is created.
RigidBodyPhysics.cpp listens to static/dynamic collision event and resolves the physics of all collisions between static and dynamic colliders (eg. mario colliding with a tile). This is all fully functional.

I fixed the order of exectution of systems (in Level.cpp) so that each collision is resolved before an entity is rendered.
TransformComponent holds pos and new_pos Vector2<int> members. pos is the position rendered for the current iteration of game loop, new_pos is what will be rendered in the next iteration, after handling of events and camera recentering on player. new_pos is calculated only once inside of RigidBodyPhysics using velocity, now all other systems (other than Camera&&RigidBodyPhysics) manipulate velocity rather than pos/new_pos to change position of entity.

Started thinking of how I could make each system hold a std::set of entities rather than a std::vector, so that the order of entities can be determined by a Comparator function provided by each System (the default would be std::less<EntityHandle>). This would allow RenderingSystem to order its entities by render layer by simply providing a Comparator function (at the moment, with std::vector, it achieves this by overriding add() and sorting its entities in there. this is kind of a hack, id rather abstract this feature).  Didn't get it to work, resigned to keeping std::vector for now.

Fixed animationcomponent and spritecomponent of level editor (ie. ImGui<AnimationComponent> and ImGui<SpriteComponent> in ImguiComponents.h) so that ImGui::Combo displays the correct animation/sprite/filepath option before ever triggering the if(ImGui::Combo...) statement. Added ImGui<RigidBodyComponent> and ImGui<ColliderComponent> as well. level editor is working well, maybe I could eventually add a prefab editor....?